+++ src/snowflake/cli/_plugins/spcs/compute_pool/manager.py
@@ -16,12 +16,17 @@ from __future__ import annotations
 
 from typing import List, Optional
 
+from snowflake.cli._plugins.object.manager import ObjectManager
 from snowflake.cli._plugins.spcs.common import (
     NoPropertiesProvidedError,
     handle_object_already_exists,
     strip_empty_lines,
 )
+from snowflake.cli._plugins.spcs.compute_pool.compute_pool_entity_model import (
+    ComputePoolEntityModel,
+)
 from snowflake.cli.api.constants import ObjectType
+from snowflake.cli.api.identifiers import FQN
 from snowflake.cli.api.sql_execution import SqlExecutionMixin
 from snowflake.connector.cursor import SnowflakeCursor
 from snowflake.connector.errors import ProgrammingError
@@ -39,7 +44,17 @@ class ComputePoolManager(SqlExecutionMixin):
         auto_suspend_secs: int,
         comment: Optional[str],
         if_not_exists: bool,
+        replace: bool,
     ) -> SnowflakeCursor:
+
+        if replace:
+            object_manager = ObjectManager()
+            object_type = ObjectType.COMPUTE_POOL.value.cli_name
+            entity_id_fqn = FQN.from_string(pool_name)
+            if object_manager.object_exists(object_type=object_type, fqn=entity_id_fqn):
+                self.stop(pool_name)
+                object_manager.drop(object_type=object_type, fqn=entity_id_fqn)
+
         create_statement = "CREATE COMPUTE POOL"
         if if_not_exists:
             create_statement = f"{create_statement} IF NOT EXISTS"
@@ -58,7 +73,25 @@ class ComputePoolManager(SqlExecutionMixin):
         try:
             return self.execute_query(strip_empty_lines(query))
         except ProgrammingError as e:
-            handle_object_already_exists(e, ObjectType.COMPUTE_POOL, pool_name)
+            handle_object_already_exists(
+                e, ObjectType.COMPUTE_POOL, pool_name, replace_available=True
+            )
+
+    def deploy(
+        self, compute_pool: ComputePoolEntityModel, replace: bool
+    ) -> SnowflakeCursor:
+        return self.create(
+            pool_name=compute_pool.entity_id,
+            min_nodes=compute_pool.min_nodes,
+            max_nodes=compute_pool.max_nodes,
+            instance_family=compute_pool.instance_family,
+            auto_resume=compute_pool.auto_resume,
+            initially_suspended=compute_pool.initially_suspended,
+            auto_suspend_secs=compute_pool.auto_suspend_seconds,
+            comment=compute_pool.comment,
+            if_not_exists=False,
+            replace=replace,
+        )
 
     def stop(self, pool_name: str) -> SnowflakeCursor:
         return self.execute_query(f"alter compute pool {pool_name} stop all")