+++ tests_integration/spcs/testing_utils/compute_pool_utils.py
@@ -18,12 +18,14 @@ import time
 import pytest
 from snowflake.connector import SnowflakeConnection
 
+from snowflake.cli.api.output.types import CommandResult
 from tests_integration.conftest import SnowCLIRunner
 from tests_integration.test_utils import contains_row_with, not_contains_row_with
 from tests_integration.testing_utils.assertions.test_result_assertions import (
     assert_that_result_is_successful_and_executed_successfully,
     assert_that_result_is_successful_and_output_json_contains,
     assert_that_result_is_successful_and_output_json_equals,
+    assert_that_result_failed_with_message_containing,
 )
 
 
@@ -62,6 +64,34 @@ class ComputePoolTestSteps:
             or f"Compute pool {compute_pool_name.upper()} successfully created."
             in result.json["status"]  # type: ignore
         )
+
+    def deploy_compute_pool(self, compute_pool_name: str) -> None:
+        result = self._deploy_compute_pool(replace=False)
+        assert_that_result_is_successful_and_output_json_contains(
+            result,
+            {"message": f"Compute pool '{compute_pool_name}' successfully deployed."},
+        )
+
+    def deploy_compute_pool_with_replace(self, compute_pool_name: str) -> None:
+        result = self._deploy_compute_pool(replace=True)
+        assert_that_result_is_successful_and_output_json_contains(
+            result,
+            {"message": f"Compute pool '{compute_pool_name}' successfully deployed."},
+        )
+
+    def second_deploy_should_fail(self) -> None:
+        result = self._deploy_compute_pool(replace=False)
+        assert_that_result_failed_with_message_containing(result, "already exists")
+
+    def _deploy_compute_pool(self, replace: bool) -> CommandResult:
+        params = [
+            "spcs",
+            "compute-pool",
+            "deploy",
+        ]
+        if replace:
+            params.append("--replace")
+        return self._setup.runner.invoke_with_connection_json(params)
 
     def list_should_return_compute_pool(self, compute_pool_name) -> None:
         result = self._execute_list()